# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登录"
                ],
                "summary": "登录接口",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.ApiResult"
                        }
                    }
                }
            }
        },
        "/api/room": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "房间相关"
                ],
                "summary": "查询房间列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.ApiResult"
                        }
                    }
                }
            }
        },
        "/api/room/exit/{rid}": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "房间相关"
                ],
                "summary": "退出房间",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "房间id",
                        "name": "rid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.ApiResult"
                        }
                    }
                }
            }
        },
        "/api/room/new": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "房间相关"
                ],
                "summary": "创建房间",
                "parameters": [
                    {
                        "description": "创建房间body",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NewRoomBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.ApiResult"
                        }
                    }
                }
            }
        },
        "/api/room/{rid}/join": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "房间相关"
                ],
                "summary": "加入房间",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "房间id",
                        "name": "rid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.ApiResult"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.NewRoomBody": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                }
            }
        },
        "res.ApiResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}